[FEAT] 열람 데이트 코스 전체 조회 API  구현 - #26

+++ dateroad-domain/src/main/java/org/dateroad/place/domain/Place.java
@@ -30,7 +30,7 @@ public abstract class Place extends BaseTimeEntity {
 
     @Column(name = "duration")
     @NotNull
-    private int duration;
+    private float duration;
 
     @Column(name = "sequence")
     @NotNull
